refactor: move cron receipt inspection into test support
